made_real/requirements.txt
absl-py==1.2.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blessings==1.7
box-overlaps==0.0.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.37.1
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.49.1
h5py==3.8.0
icecream==2.1.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipdb==0.13.9
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyter_core==4.12.0
jupyter_packaging==0.12.3
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
jupyterlab_server==2.15.1
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.6
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-full==1.6.1
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.4
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nuscenes-devkit==1.1.10
nvidia-ml-py==11.495.46
nvidia-ml-py3==7.352.0
oauthlib==3.2.1
omegaconf==2.2.3
open3d==0.15.2
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pypcd==0.1.1
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-lzf==0.2.4
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
Shapely==1.8.4
sniffio==1.2.0
sort-vertices==0.0.0
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.15.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.11.4
torch==1.8.1
torch-tb-profiler==0.4.0
torchattacks==3.2.7
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tornado==6.2
tqdm
traitlets==5.9.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
widgetsnbextension==4.0.2
wrapt==1.15.0
yapf==0.32.0
zipp==3.8.1
